fix: attach cloud event to context as a JS object
This change is quite simple, really - just calling .format()
on the cloud event. In order to test it, however, I made a
few changes. The cloud event unmarshalling happens in its
own module now. The module exports an async function to do
the unmarshalling, and the main request handler returns its result
to the caller.