doc: use oscf/js-runtime in running locally
This commit uses the image name built by the make build target, instead
of redhat-faas/js-runtime.

It also removes the cidfile as this can be confusing to new users as
unless the manually delete this file there will be an error the next
time it is run:

docker: Container ID file found, make sure the other container isn't
running or delete my-faas-test.cid.