src: change to use faas-js-runtime v0.0.2
The 0.0.2 version of the runtime does not install dependencies
when the framework spins up. That is the responsibility of the
user. In our case, the assemble script does the npm install for
any user functions. The application entry point at index.js now
passes a function instead of a path to a function to the framework.