src: extract runtime framework into its own module (#29)
The runtime framework can be developed tested and published
independent of the runtime image by pulling the source out
of this repository and publishing it in the npm repository
at https://www.npmjs.com/package/@redhat/faas-js-runtime.

The framework source has been copied to another repository
at https://github.com/openshift-cloud-functions/faas-js-runtime.

Fixes: https://github.com/openshift-cloud-functions/faas-js-runtime-image/issues/28

build: bump image tag to 0.0.2

The 0.0.2 version of the runtime does not install dependencies
when the framework spins up. That is the responsibility of the
user. In our case, the assemble script does the npm install for
any user functions. The application entry point at index.js now
passes a function instead of a path to a function to the framework.